% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradable_industry_list.R
\name{degradable_industry_list}
\alias{degradable_industry_list}
\title{This function will provide an output list of industries that can be used in the coming analysis. This should be used to help determine where different shocks should/can be applied.}
\usage{
degradable_industry_list(io_data)
}
\arguments{
\item{io_data}{An io_data list object, generated as the output of the **draw_io_data** function}
}
\value{
A list of the industries involved in the analysis
}
\description{
This function will provide an output list of industries that can be used in the coming analysis. This should be used to help determine where different shocks should/can be applied.
}
