% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade_industries.R
\name{degrade_industries}
\alias{degrade_industries}
\title{This function will allow the creation of a dataframe that will be used as an input for the iim/diim functions (or similar). Essentially, it will show a list of industries and the applicable shocks to final demand.}
\usage{
degrade_industries(
  io_data,
  degraded_industry_single_name,
  degraded_industry_single_shock = 0.1,
  degraded_industry_dataframe = NULL
)
}
\arguments{
\item{io_data}{An io_data list object, generated as the output of the **draw_io_data** function}

\item{degraded_industry_single_name}{The name of an industry to be shocked as part of the IO analysis}

\item{degraded_industry_single_shock}{A proportion of final demand to be shocked as part of the IO analysis}

\item{degraded_industry_dataframe}{Where more than one industry should experience a shock (for example, when analysing a natural disaster), a dataframe of shocks can be passed here}
}
\value{
A dataframe of degraded final demand where the Degradation_Proportion or Degradation_Absolute columns can be used in the respective IO analyses
}
\description{
This function will allow the creation of a dataframe that will be used as an input for the iim/diim functions (or similar). Essentially, it will show a list of industries and the applicable shocks to final demand.
}
