% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diim.R
\name{diim}
\alias{diim}
\title{This function will perform a dynamic inoperability input-output analysis on an economy}
\usage{
diim(
  totaloutput,
  finaldemand,
  interindustry,
  industrylist,
  degradation,
  degradationtype = "Proportion",
  type = 1,
  estimated_recoveryduration,
  estimated_initialinoperability,
  estimated_finalinoperability,
  analysisperiod
)
}
\arguments{
\item{totaloutput}{The total output vector, most easily collected from the **draw_io_data** function (io_data$TotalOutput)}

\item{finaldemand}{The final demand vector, most easily collected from the **draw_io_data** function (io_data$FinalDemand)}

\item{interindustry}{The inter-industry matrix, most easily collected from the **draw_io_data** function (io_data$InterIndustry)}

\item{industrylist}{The industry list, most easily collected from the **draw_io_data** function (io_data$Industries)}

\item{degradation}{The degradation vector, most easily collected from the **degrade_industries** function (degradation$Degradation_Proportion)}

\item{degradationtype}{The degradation approach ("Propertion" or "Absolute")}

\item{type}{The IO analysis type where type 1 is not including households as a sector and type 2 (not yet available) where households are included as a sector}

\item{estimated_recoveryduration}{The number of periods expected to generate a recovery from the inital to final inoperability measures}

\item{estimated_initialinoperability}{The initial inoperability assumed as part of the recovery}

\item{estimated_finalinoperability}{The final inoperability assumed as part of the recovery (typically very, very small - like 0.001)}

\item{analysisperiod}{The number of periods over which the DIIM should be conducted}
}
\value{
A dataframe including the inoperability and total economic loss for each sector in each time period after the shock
}
\description{
This function will perform a dynamic inoperability input-output analysis on an economy
}
