% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_io_data.R
\name{draw_io_data}
\alias{draw_io_data}
\title{This function will perform a mass collection of IO data from a particular system}
\usage{
draw_io_data(
  data,
  full_range,
  Z_range,
  C_range,
  I_range,
  G_range,
  LAB_range,
  GS_range,
  X_range,
  NRH_range,
  f_range,
  industries_range
)
}
\arguments{
\item{data}{The filepath to the IO Excel file}

\item{full_range}{The cell range over which the full outputs from production for each sector are calculated}

\item{Z_range}{The range over which the Z matrix is collected - ie, the inter-industry data. Should be a square matrix}

\item{C_range}{The range over which the C matrix is collected - ie, the consumption data for each industry}

\item{I_range}{The range over which the I matrix is collected - ie, the investment data for each industry}

\item{G_range}{The range over which the G matrix is collected - ie, the government spending data for each industry}

\item{LAB_range}{The range over which the LAB matrix is collected}

\item{GS_range}{The range over which the GS matrix is collected}

\item{X_range}{The range over which the X matrix is collected - ie, the export data for each industry}

\item{NRH_range}{The range over which the NRH matrix is collected - ie, the non-residential household consumption data for each industry}

\item{f_range}{The range over which the f matrix is collected - ie, the final demand for each industry}

\item{industries_range}{The range over which the industry matrix is collected - ie, the list of industries}
}
\value{
A large list of all the individual matrix elements required for an IO analysis
}
\description{
This function will perform a mass collection of IO data from a particular system
}
