% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iim.R
\name{iim}
\alias{iim}
\title{This function will perform an inoperability input-output analysis on an economy}
\usage{
iim(
  totaloutput,
  finaldemand,
  interindustry,
  industrylist,
  degradation,
  degradationtype = "Proportion",
  type = 1
)
}
\arguments{
\item{totaloutput}{The total output vector, most easily collected from the **draw_io_data** function (io_data$TotalOutput)}

\item{finaldemand}{The final demand vector, most easily collected from the **draw_io_data** function (io_data$FinalDemand)}

\item{interindustry}{The inter-industry matrix, most easily collected from the **draw_io_data** function (io_data$InterIndustry)}

\item{industrylist}{The industry list, most easily collected from the **draw_io_data** function (io_data$Industries)}

\item{degradation}{The degradation vector, most easily collected from the **degrade_industries** function (degradation$Degradation_Proportion)}

\item{degradationtype}{The degradation approach ("Propertion" or "Absolute")}

\item{type}{The IO analysis type where type 1 is not including households as a sector and type 2 (not yet available) where households are included as a sector}
}
\value{
A dataframe including the inoperability and total economic loss for each sector
}
\description{
This function will perform an inoperability input-output analysis on an economy
}
