% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_degradation.R
\name{stochastic_degradation}
\alias{stochastic_degradation}
\title{This function will take the degradation table created in the **degrade_industries** function and add random noise to the degradation values}
\usage{
stochastic_degradation(
  degradation,
  stochastic_modifier = NULL,
  stochastic_modifier_arg1 = NULL,
  stochastic_modifier_arg2 = NULL,
  stochastic_modifier_arg3 = NULL
)
}
\arguments{
\item{degradation}{A degradation table output of the **degrade_industries** function}

\item{stochastic_modifier}{A stochastic model to be used to add random noise to the degradation: *currently only supports the Normal distribution (rnorm)*}

\item{stochastic_modifier_arg1}{A parameter for the stochastic modifier}

\item{stochastic_modifier_arg2}{A parameter for the stochastic modifier}

\item{stochastic_modifier_arg3}{A parameter for the stochastic modifier}
}
\value{
An updated dataframe of the inital shocks to final demand, with the random noise added
}
\description{
This function will take the degradation table created in the **degrade_industries** function and add random noise to the degradation values
}
