% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_collect_structures.R
\name{collect_structures}
\alias{collect_structures}
\title{This function ultimately will collect data from an IO structure}
\arguments{
\item{file}{The root IO file}

\item{range}{The Excel range over which the data are held}

\item{numeric}{This determines whether the data are number, automatically set to TRUE/1}
}
\value{
A dataframe of data collected from the Excel file
}
\description{
This function ultimately will collect data from an IO structure
}
