% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{combine_structures}
\alias{combine_structures}
\title{This function will combine previously collected data structures to ensure only a single column}
\usage{
combine_structures(data, column = T)
}
\arguments{
\item{data}{A dataframe that should be summed together to create a singular vector}

\item{column}{A boolean determining whether the data should be aggregated into a single column (or alternatively, row)}
}
\value{
A singular column or row of totals, based on the inputs
}
\description{
This function will combine previously collected data structures to ensure only a single column
}
